+++ Wrappers/InvoiceWrapper.cs
@@ -70,7 +70,7 @@ namespace Facturapi.Wrappers
             return invoice;
         }
 
-        public async Task SendByEmailAsync(string id, Dictionary<string, object> data)
+        public async Task SendByEmailAsync(string id, Dictionary<string, object> data = null)
         {
             var response = await client.PostAsync(Router.SendByEmail(id), new StringContent(JsonConvert.SerializeObject(data), Encoding.UTF8, "application/json"));
             if (!response.IsSuccessStatusCode)
@@ -182,6 +182,19 @@ namespace Facturapi.Wrappers
             }
             var invoice = JsonConvert.DeserializeObject<Invoice>(resultString, this.jsonSettings);
             return invoice;
+        }
+
+        public async Task<Stream> PreviewPdfAsync(Dictionary<string, object> query = null)
+        {
+            var response = await client.GetAsync(Router.PreviewPdf(query));
+            if (!response.IsSuccessStatusCode)
+            {
+                var resultString = await response.Content.ReadAsStringAsync();
+                var error = JsonConvert.DeserializeObject<JObject>(resultString, this.jsonSettings);
+                throw new FacturapiException(error["message"].ToString());
+            }
+            var stream = await response.Content.ReadAsStreamAsync();
+            return stream;
         }
     }
 }