+++ CHANGELOG.md
@@ -5,14 +5,52 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.7.1] = 2025-10-23
+## [4.10.1] = 2025-10-23
 
 ### Added
 
 - Added fields to `Customer` model: `DefaultInvoiceUse`.
 
-## [4.6.0] - 2024-23-09
+## [4.10.0] - 2025-09-04
 
+### Added
+
+- Add `Invoices.PreviewPdfAsync` method to preview invoice PDF before stamping.
+
+## [4.9.1] - 2025-07-21
+
+### Fixed
+
+- Fix `Invoices.SendByEmailAsync`, `Receipts.SendByEmailAsync`, `Retentions.SendByEmailAsync` method and mark the `data` parameter as optional.
+
+### Added
+
+- Add `AddOns` property to `Organization` model to support Facturapi add-ons.
+- Add missing properties to `Organization.Customization.PdfExtra` model: `AddressCodes`, `RoundUnitPrice`, `TaxBreakdown`, `IepsBreakdown`, `RenderCartaPorte`, and `RepeatSignature`.
+
+## [4.9.0] - 2025-06-16
+
+### Added
+
+- Add `Organizations.UpdateSelfInvoiceSettingsAsync` method to update self invoice settings.
+
+## [4.8.0] - 2025-04-22
+
+### Added
+
+- Add Create Webhook `Webhooks.CreateAsync`
+- Add Update Webhook `Webhooks.UpdateAsync`
+- Add Retrieve Webhook `Webhooks.RetrieveAsync`
+- Add Delete Webhook `Webhooks.DeleteAsync`
+- Add List Webhooks `Webhooks.ListAsync`
+- Add Validate Signature Webhook `Webhooks.ValidateSignatureAsync`
+
+## [4.7.1] - 2025-04-16
+
+### Added
+
+- Type IepsMode for Tax model 
+- Type Factor for Tax model
 
 ## [4.7.0] = 2025-02-25
 
@@ -22,14 +60,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added fields to `Customer` model: `SatValidatedAt`, `EditLink` and `EditLinkExpiresAt`.
 - Added targets for .NET 6.0 and .NET 7.0.
 
-## [4.6.0] - 2024-23-09
+## [4.6.0] - 2024-09-23
 
 ### Added
 
 - Add List of Live Api Keys `Organizations.ListAsyncLiveApiKey`
 - Add Delete of a Live Api Key `Organization.DeleteAsyncLiveApiKey`
 
-## [4.5.0] - 2024-06-05
+
+## [4.5.0] - 2024-05-06
 
 ### Added